MADCFusion/requirements.txt
absl-py==1.0.0
addict==2.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
bleach==6.1.0
brotlipy==0.7.0
cachetools==5.0.0
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==3.1.0
colorama==0.4.4
contourpy==1.0.6
cryptography==3.4.8
cycler==0.11.0
Cython==0.29.32
dataclasses==0.8
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
einops==0.6.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.10.7
flit_core==3.7.1
fonttools==4.33.3
fsspec==2024.3.1
ftfy==6.1.1
future==0.18.2
google-auth==2.17.1
google-auth-oauthlib==1.0.0
grpcio==1.53.0
h5py==3.7.0
huggingface-hub==0.13.3
idna==3.4
imageio==2.19.5
imgaug==0.4.0
importlib-metadata==4.11.4
importlib-resources==5.7.1
ipykernel==6.14.0
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.0.2
IQA-pytorch==0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter_core==5.7.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.2
kornia==0.6.7
lmdb==1.4.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-service==2.4.0
mpmath==1.3.0
munkres==1.1.4
natsort==8.3.1
nb-conda==2.2.1
nb-conda-kernels==2.3.1
nbclient==0.6.4
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.5
networkx==2.8.5
nibabel==4.0.2
notebook==6.4.12
numexpr==2.8.4
numpy==1.22.3
oauthlib==3.2.0
openai-clip==1.0.1
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
pip==22.1.2
pipreqs==0.5.0
platformdirs==4.2.2
prometheus-client==0.14.1
prompt_toolkit==3.0.47
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==0.18.2
Pygments==2.12.0
pyiqa==0.1.6.3
pyOpenSSL==20.0.1
pyparsing==3.0.8
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.1.0
qqdm==0.0.7
qtconsole==5.3.2
QtPy==2.2.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.8
runstats==2.0.0
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.10.1
Send2Trash==1.8.0
setuptools==65.3.0
shapely==2.0.1
six==1.16.0
some-package==0.1
soupsieve==2.3.1
stack-data==0.3.0
sympy==1.12
tables==3.7.0
tb-nightly==2.13.0a20230330
tensorboard==2.9.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
terminado==0.15.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.6.13
tinycss2==1.1.1
torch==2.2.1
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.13.1a0
tornado==6.1
tqdm==4.64.0
traitlets==5.14.3
typing_extensions==4.10.0
unicodedata2==14.0.0
urllib3==1.26.15
ViTables==3.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
wheel==0.37.1
widgetsnbextension==4.0.3
win-inet-pton==1.1.0
xlrd==2.0.1
yapf==0.32.0
yarg==0.1.9
zipp==3.8.0
